entity_code_redesign
Changes architecture of EntityCombatant and adds EntityHuman, which is identical to EntityCombatant save for the fact that it has keybindings.  The control flow of an entities actions has been broken out in the following way:  Both entities have methods that correspond to actions, those methods govern properties that describe the nature of the actions (e.g. walkingDirection: 'left'), the update function automatically makes the entity 'go' based on the properties it needs to observe.  This PR also implements a working prototype of the multiplayer server, and adds socket messaging support for EntityHuman and EntityCombatant.