multiplayer_server_debug
This commit ammends some of the wacky socket behavior that was being experienced following the initial implementation of the multiplayer server and the entity re-design.  Players now appear in their proper place in the world for they are joining or being joined (in other words, if you join a game you can now see pre-existing players), When a player dies or disconnects their entity will now be removed from the game.