level_polish
This commit sets the stage for the game formats of single player and multiplayer co-op by implementing an elongated bounding box and adding a background graphic.  More importantly, this commit dispenses with the custom movement trace that was being used to handle jumping through platforms and leaves that job to weltmeister.