evil_clone_entity
This commit adds the evil clone entity as well as the hitbox entity which is utilized (for now) only by the human entity.  Clones do damage on contact and humans do damage through strikes.  No animations or health bars reflect this right now, though.  Also, Clones are spawned in Weltmeister, (for continued testing purposes), and are not persisted across the network.